update issue details to entity issue details
