+++ src/controllers/OrganisationsController.js
@@ -1,6 +1,6 @@
 import datasetTaskListMiddleware from '../middleware/datasetTaskList.middleware.js'
 import datasetOverviewMiddleware from '../middleware/datasetOverview.middleware.js'
-import issueDetailsMiddleware from '../middleware/issueDetails.middleware.js'
+import entityIssueDetailsMiddleware from '../middleware/entityIssueDetails.middleware.js'
 import issueTableMiddleware from '../middleware/issueTable.middleware.js'
 import organisationsMiddleware from '../middleware/organisations.middleware.js'
 import getStartedMiddleware from '../middleware/getStarted.middleware.js'
@@ -11,7 +11,7 @@ const organisationsController = {
   organisationsMiddleware,
   datasetTaskListMiddleware,
   datasetOverviewMiddleware,
-  issueDetailsMiddleware,
+  entityIssueDetailsMiddleware,
   issueTableMiddleware,
   getStartedMiddleware,
   overviewMiddleware,
