+++ src/routes/organisations.js
@@ -7,8 +7,8 @@ router.get('/:lpa/:dataset/get-started', OrganisationsController.getStartedMiddl
 router.get('/:lpa/:dataset/overview', OrganisationsController.datasetOverviewMiddleware)
 router.get('/:lpa/:dataset/data/:pageNumber', OrganisationsController.datasetDataviewMiddleware)
 router.get('/:lpa/:dataset/data', OrganisationsController.datasetDataviewMiddleware)
-router.get('/:lpa/:dataset/:issue_type/:issue_field/entry/:pageNumber', OrganisationsController.issueDetailsMiddleware)
-router.get('/:lpa/:dataset/:issue_type/:issue_field/entry', OrganisationsController.issueDetailsMiddleware)
+router.get('/:lpa/:dataset/:issue_type/:issue_field/entity/:pageNumber', OrganisationsController.entityIssueDetailsMiddleware)
+router.get('/:lpa/:dataset/:issue_type/:issue_field/entity', OrganisationsController.entityIssueDetailsMiddleware)
 router.get('/:lpa/:dataset/:issue_type/:issue_field/:pageNumber?', OrganisationsController.issueTableMiddleware)
 router.get('/:lpa/:dataset/:issue_type/:issue_field', OrganisationsController.issueTableMiddleware)
 router.get('/:lpa/:dataset', OrganisationsController.datasetTaskListMiddleware)
